+++ src/main/java/com/twilio/rest/verify/v2/Service.java
@@ -32,13 +32,9 @@ import java.net.URI;
 import java.util.Map;
 import java.util.Objects;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to
- * change. Use them with caution.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Service extends Resource {
-    private static final long serialVersionUID = 102125426973308L;
+    private static final long serialVersionUID = 28072870212695L;
 
     /**
      * Create a ServiceCreator to execute create.