+++ src/main/java/com/twilio/rest/verify/v2/service/RateLimit.java
@@ -32,10 +32,6 @@ import java.net.URI;
 import java.util.Map;
 import java.util.Objects;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to
- * change. Use them with caution.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class RateLimit extends Resource {
     private static final long serialVersionUID = 110366061579267L;