+++ src/main/java/com/twilio/rest/verify/v2/service/Verification.java
@@ -34,10 +34,6 @@ import java.net.URI;
 import java.util.Map;
 import java.util.Objects;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to
- * change. Use them with caution.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Verification extends Resource {
     private static final long serialVersionUID = 90062652968318L;