+++ src/main/java/com/twilio/rest/verify/v2/service/VerificationCheckCreator.java
@@ -17,10 +17,6 @@ import com.twilio.http.Response;
 import com.twilio.http.TwilioRestClient;
 import com.twilio.rest.Domains;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to
- * change. Use them with caution.
- */
 public class VerificationCheckCreator extends Creator<VerificationCheck> {
     private final String pathServiceSid;
     private final String code;
@@ -43,11 +39,12 @@ public class VerificationCheckCreator extends Creator<VerificationCheck> {
     }
 
     /**
-     * The phone number to verify. Either this parameter or the `verification_sid`
-     * must be specified. Phone numbers must be in [E.164
-     * format](https://www.twilio.com/docs/glossary/what-e164)..
+     * The phone number or
+     * [email](https://www.twilio.com/docs/verify/tutorials/email) to verify. Either
+     * this parameter or the `verification_sid` must be specified. Phone numbers
+     * must be in [E.164 format](https://www.twilio.com/docs/glossary/what-e164)..
      *
-     * @param to The phone number to verify
+     * @param to The phone number or email to verify
      * @return this
      */
     public VerificationCheckCreator setTo(final String to) {
@@ -57,7 +54,9 @@ public class VerificationCheckCreator extends Creator<VerificationCheck> {
 
     /**
      * A SID that uniquely identifies the Verification Check. Either this parameter
-     * or the `to` phone number must be specified..
+     * or the `to` phone
+     * number/[email](https://www.twilio.com/docs/verify/tutorials/email) must be
+     * specified..
      *
      * @param verificationSid A SID that uniquely identifies the Verification Check
      * @return this